requirements.txt
asgiref==3.7.2
Django==3.2.5
djangorestframework==3.14.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
